+++ src/hs485d/HS485CommMessage.h
@@ -39,7 +39,7 @@ public:
 	virtual void SetTimeout(int to);
 	virtual bool TransformToSimulationMessage();
 
-	virtual bool MatchType(unsigned long type);
+	virtual bool MatchType(uint32_t type);
 	virtual uint32_t GetSenderAddress();
 	virtual uint32_t GetReceiverAddress();
 	virtual void SetSenderAddress(uint32_t address);
@@ -62,9 +62,9 @@ protected:
 	bool eepRomUsage;
 
 	virtual int MapIndex(int index);
-    virtual unsigned long GetType();
+  virtual uint32_t GetType();
 	virtual bool ProcessResponse(CommMessage* m, t_state* new_state);
-	virtual void SetType(unsigned long type);
+	virtual void SetType(uint32_t type);
 	friend class HS485CommMessageDecoder;
 };
 #endif //_HS485_COMM_MESSAGE_H_
